made a tab bar for reviews screen and embedded reviewed container and to review container in it
