+++ itour_planner/lib/ReviewsScreen/ReviewedScreen/reviewed_container.dart
@@ -22,7 +22,6 @@ class ReviewContainer extends StatelessWidget {
     return GestureDetector(
       onTap: () {},
       child: Container(
-        margin: const EdgeInsets.all(5),
         padding: const EdgeInsets.only(bottom: 20),
         width: double.infinity,
         child: Column(
